Use Instance object as interface between internal and external heap
