+++ src/core/Makefile.am
@@ -225,8 +225,7 @@ SOURCES_COMMON                             = \
     utils/channel_manager.cpp                \
     utils/channel_monitor.cpp                \
     utils/child_supervision.cpp              \
-    utils/external_heap.cpp                  \
-    utils/internal_heap.cpp                  \
+    utils/heap.cpp                           \
     utils/jam_detector.cpp                   \
     utils/missing_strlcat.c                  \
     utils/missing_strlcpy.c                  \
@@ -429,8 +428,6 @@ HEADERS_COMMON                             = \
     utils/channel_manager.hpp                \
     utils/channel_monitor.hpp                \
     utils/child_supervision.hpp              \
-    utils/external_heap.hpp                  \
-    utils/internal_heap.hpp                  \
     utils/heap.hpp                           \
     utils/jam_detector.hpp                   \
     utils/parse_cmdline.hpp                  \
