+++ src/core/crypto/mbedtls.cpp
@@ -53,12 +53,12 @@ namespace Crypto {
 
 static void *CAlloc(size_t aCount, size_t aSize)
 {
-    return Instance::Get().GetHeap().CAlloc(aCount, aSize);
+    return Instance::Get().HeapCAlloc(aCount, aSize);
 }
 
 static void Free(void *aPointer)
 {
-    Instance::Get().GetHeap().Free(aPointer);
+    Instance::Get().HeapFree(aPointer);
 }
 
 #endif // !OPENTHREAD_CONFIG_MULTIPLE_INSTANCE_ENABLE && OPENTHREAD_CONFIG_ENABLE_BUILTIN_MBEDTLS
